+++ ExilePartyClient/src/app/shared/providers/external.service.ts
@@ -17,6 +17,7 @@ import { Stash } from '../interfaces/stash.interface';
 import { AnalyticsService } from './analytics.service';
 import { ElectronService } from './electron.service';
 import { LogService } from './log.service';
+import { Character } from '../interfaces/character.interface';
 
 @Injectable()
 export class ExternalService {
@@ -44,18 +45,6 @@ export class ExternalService {
     return this.http.get('https://www.pathofexile.com/character-window/get-items' + parameters, { withCredentials: true }).catch(e => {
       if (e.status !== 403 && e.status !== 404) {
         this.logService.log('Could not character items, disconnecting!', null, true);
-        this.router.navigate(['/disconnected']);
-      }
-      return Observable.of(null);
-    });
-  }
-
-  getLeagues(): Observable<any> {
-    const leagueType = 'main';
-    const parameters = `?type=${leagueType}`;
-    return this.http.get('http://api.pathofexile.com/leagues' + parameters).catch(e => {
-      if (e.status !== 403 && e.status !== 404) {
-        this.logService.log('Could not fetch leagues, disconnecting!', null, true);
         this.router.navigate(['/disconnected']);
       }
       return Observable.of(null);